Fixed page structure and home links
