+++ 11. Trivia/src/pages/blizzard-trivia.html
@@ -32,7 +32,6 @@
     >
       <div id="navbar"></div>
     </div>
-
     <div
       class="container mx-auto mt-2 md:mt-4 p-2 md:pb-5 md:px-5 rounded-lg border shadow-xl"
     >
@@ -46,7 +45,7 @@
       id="footer"
       class="container mx-auto mt-2 md:mt-3 p-2 md:py-5 md:px-5 rounded-lg border shadow-xl"
     ></div>
-    <script type="module" src="./blizzard-trivia.js"></script>
+    <script type="module" src="../scripts/blizzard-trivia.js"></script>
     <script type="module" src="../scripts/blizzardTriviaGame.js"></script>
   </body>
 </html>
