+++ 11. Trivia/src/pages/index.html
@@ -20,7 +20,7 @@
       href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="./style.css" />
+    <link rel="stylesheet" href="../style.css" />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.9.0/css/all.min.css"
@@ -46,6 +46,6 @@
       id="footer"
       class="container mx-auto mt-2 md:mt-3 p-2 md:py-5 md:px-5 rounded-lg border shadow-xl"
     ></div>
-    <script type="module" src="./app.js"></script>
+    <script type="module" src="../scripts/app.js"></script>
   </body>
 </html>
