+++ 11. Trivia/src/scripts/blizzard-trivia.js
@@ -1,7 +1,19 @@
+import questions from "../data/questions.json" assert { type: "json" };
+import answers from "../data/answers.json" assert { type: "json" };
+
 const navbar = document.querySelector("#navbar");
 const appHead = document.querySelector("#app-head");
 const appContent = document.querySelector("#app-content");
 const footer = document.querySelector("#footer");
+
+const blizzTitle = "Blizzard";
+let questionId = 0;
+let answerID = 0;
+let blizzQuestion = questions.blizzardQuestions[questionId].question;
+let blizzAnswerOneText = answers.blizzardAnswers[answerID].answerOne;
+let blizzAnswerTwoText = answers.blizzardAnswers[answerID].answertwo;
+let blizzAnswerThreeText = answers.blizzardAnswers[answerID].answerThree;
+let blizzAnswerFourText = answers.blizzardAnswers[answerID].answerFour;
 
 // App Markup
 const navbarContent = `
@@ -28,7 +40,7 @@ const navbarContent = `
                 tabindex="0"
                 class="menu menu-compact dropdown-content mt-3 p-2 shadow bg-base-100 rounded-box w-52"
             >
-                <li><a class="hover:text-sky-400  hover:border-b hover:border-orange-400" href="./index.html">Home</a></li>
+                <li><a class="hover:text-sky-400  hover:border-b hover:border-orange-400" href="index.html">Home</a></li>
                 <li>
                 <a
                     class="hover:text-sky-400  hover:border-b hover:border-sky-400"
@@ -54,7 +66,7 @@ const navbarContent = `
         <li
             class="hover:border-b hover:border-sky-400"
         >
-            <a href="./index.html"><span class="text-sky-400 hover:text-orange-400">Home</span></a>
+            <a href="index.html"><span class="text-sky-400 hover:text-orange-400">Home</span></a>
         </li>
         <li class="hover:text-orange-400 hover:border-b hover:border-sky-400">
             <a
@@ -88,13 +100,13 @@ const appBodyContent = `
 <div class="card lg:card-side bg-base-100 shadow-xl">
   <figure><img src="https://placeimg.com/400/400/tech" alt="Album"></figure>
   <div class="card-body">
-    <h2 class="card-title">Trivia Game!</h2>
-    <p>Welcome to the trivia game! We currently have four themed trivia modules. Choose one below!</p>
-    <div class="card-actions justify-center md:justify-end">
-      <button id="btn-blizzard" class="btn btn-primary md:btn-outline">Blizzard</button>
-      <button id="btn-warcraft" class="btn btn-success md:btn-outline">Warcraft</button>
-      <button id="btn-starcraft" class="btn btn-secondary md:btn-outline">Starcraft</button>
-      <button id="btn-diablo" class="btn btn-error md:btn-outline">Diablo</button>
+    <h2 class="card-title">${blizzTitle}:</h2>
+    <p>${blizzQuestion}</p>
+    <div id="options" class="card-actions justify-center md:justify-end">
+      <button id="btn-answerOne" class="btn btn-primary md:btn-outline">${blizzAnswerOneText}</button>
+      <button id="btn-answerTwo" class="btn btn-success md:btn-outline">${blizzAnswerTwoText}</button>
+      <button id="btn-answerThree" class="btn btn-secondary md:btn-outline">${blizzAnswerThreeText}</button>
+      <button id="btn-answerFour" class="btn btn-error md:btn-outline">${blizzAnswerFourText}</button>
     </div>
   </div>
 </div>
@@ -104,16 +116,16 @@ const footerContent = `
 <footer class="footer items-center p-4 text-neutral-content rounded-md">
   <div class="items-center grid-flow-col mt-4 rounded-lg">
     <a target="_blank" class="invisible md:visible cursor-none">
-        <img src="./images/icons/snowman.png" alt="snowman" class="w-0 md:w-10 h-0 md:h-10">
+        <img src="../images/icons/snowman.png" alt="snowman" class="w-0 md:w-10 h-0 md:h-10">
     </a>
     <p>Copyright Kyle Martin © 2022 - All right reserved</p>
   </div> 
   <div class="grid-flow-col gap-4 md:place-self-center md:justify-self-end invisible md:visible">
     <a href="https://github.com/Thesnowmanndev" target="_blank">
-        <img src="./images/icons/github.png" alt="github" class="w-0 md:w-10 h-0 md:h-10">
+        <img src="../images/icons/github.png" alt="github" class="w-0 md:w-10 h-0 md:h-10">
     </a> 
     <a href="https://www.linkedin.com/in/developer-kyle-martin/" target="_blank">
-        <img src="./images/icons/linkedin.png" alt="linkedin" class="w-0 md:w-10 h-0 md:h-10">
+        <img src="../images/icons/linkedin.png" alt="linkedin" class="w-0 md:w-10 h-0 md:h-10">
     </a> 
   </div>
 </footer>
@@ -133,30 +145,3 @@ let createApp = () => {
 };
 
 createApp();
-
-// App Logic
-const btnBlizzard = document.querySelector("#btn-blizzard");
-const btnWarcraft = document.querySelector("#btn-warcraft");
-const btnStarcraft = document.querySelector("#btn-starcraft");
-const btnDiablo = document.querySelector("#btn-diablo");
-
-btnBlizzard.onclick = (event) => {
-  event.preventDefault();
-
-  window.open("../src/pages/blizzard-trivia.html", "_self");
-};
-
-btnWarcraft.onclick = (event) => {
-  event.preventDefault();
-};
-
-btnStarcraft.onclick = (event) => {
-  event.preventDefault();
-};
-
-btnDiablo.onclick = (event) => {
-  event.preventDefault();
-};
-
-// Debug
-let logClicked = () => console.log("Clicked");
