+++ 11. Trivia/src/styles/style.css
